functions/requirements.txt
firebase-functions
firebase-admin
langdetect
google-generativeai
fastapi
mangum
flask
